Move PbList to its own library (#1063)
A while ago we started converting parts to libraries to allow conditional
compilation.

This continues to move more parts to libraries, with `PbList`.

Libraries allow specifying public and internally public interfaces of the types
and makes it easier to define and maintain internal invariants.

To avoid large (difficult to sync) diffs, I'm moving one part at a time.

apitool reports no changes with this PR.