Check key and value validity in PbMap (#1076)
Currently `PbMap` does not check ranges of values added as keys or values.

This allows serializing and deserializing a message with a map and getting a
different map back, because `PbMap` allows adding values that are out of range
(e.g. an integer larger than 32-bit range as a `sfixed32` value) but when
serializing it truncates the value.

Update `PbMap` to use the same validation functions as `PbList` when adding
elements. Both keys and values are checked.

Fixes #1065.